Fix crashes when accessing null pipe
It could happen that guest will send partial command
to the host, for example, only send "acce" instead of
"accept". In that case, the connected_pipe of adb_state
will be NULL and will crash emulator.

We should use apipe instead, which is the correct one
and will be the same to connected_pipe (if connected_pipe
is not null).

Change-Id: Ic713fef8d963ba197babd4602b94edc88c64ac08
(cherry picked from commit 34069d10615f5b7c9feb4915b3d8ee62d4d98496)